// packages/types/src/interfaces/metadata/types.ts
export interface CustomMetadata15 extends Struct {
  readonly map: BTreeMap<Text, CustomValueMetadata15>;
}

/** @name CustomValueMetadata15 */
export interface CustomValueMetadata15 extends Struct {
  readonly type: SiLookupTypeId;
  readonly value: Bytes;
}

/** @name EnumDeprecationInfoV16 */
export interface EnumDeprecationInfoV16 extends BTreeMap<u8, VariantDeprecationInfoV16> {}
// ...
export interface ExtrinsicMetadataV16 extends Struct {
  readonly versions: Bytes;
  readonly addressType: SiLookupTypeId;
  readonly callType: SiLookupTypeId;
  readonly signatureType: SiLookupTypeId;
  readonly transactionExtensionsByVersion: BTreeMap<u8, Vec<Compact<u32>>>;
  readonly transactionExtensions: Vec<TransactionExtensionMetadataV16>;
}
// ...
export interface MetadataLatest extends MetadataV16 {}
// ...
export interface MetadataV16 extends Struct {
  readonly lookup: PortableRegistry;
  readonly pallets: Vec<PezpalletMetadataV16>;
  readonly extrinsic: ExtrinsicMetadataV16;
  readonly apis: Vec<RuntimeApiMetadataV16>;
  readonly outerEnums: OuterEnums15;
  readonly custom: CustomMetadata15;
}
// ...
export interface OuterEnums15 extends Struct {
  readonly callType: SiLookupTypeId;
  readonly eventType: SiLookupTypeId;
  readonly errorType: SiLookupTypeId;
}

/** @name PezpalletAssociatedTypeMetadataV16 */
export interface PezpalletAssociatedTypeMetadataV16 extends Struct {
  readonly name: Text;
  readonly type: SiLookupTypeId;
  readonly docs: Vec<Text>;
}

/** @name PezpalletCallMetadataLatest */
export interface PezpalletCallMetadataLatest extends PezpalletCallMetadataV16 {}

/** @name PezpalletCallMetadataV14 */
export interface PezpalletCallMetadataV14 extends Struct {
  readonly type: SiLookupTypeId;
}

/** @name PezpalletCallMetadataV16 */
export interface PezpalletCallMetadataV16 extends Struct {
  readonly type: SiLookupTypeId;
  readonly deprecationInfo: EnumDeprecationInfoV16;
}
// ...
export interface PezpalletMetadataV16 extends Struct {
  readonly name: Text;
  readonly storage: Option<PezpalletStorageMetadataV16>;
  readonly calls: Option<PezpalletCallMetadataV16>;
  readonly events: Option<PezpalletEventMetadataV16>;
  readonly constants: Vec<PezpalletConstantMetadataV16>;
  readonly errors: Option<PezpalletErrorMetadataV16>;
  readonly associatedTypes: Vec<PezpalletAssociatedTypeMetadataV16>;
  readonly viewFunctions: Vec<PezpalletViewFunctionMetadataV16>;
  readonly index: u8;
  readonly docs: Vec<Text>;
  readonly deprecationInfo: ItemDeprecationInfoV16;
}
// ...
export interface RuntimeApiMetadataV16 extends Struct {
  readonly name: Text;
  readonly methods: Vec<RuntimeApiMethodMetadataV16>;
  readonly docs: Vec<Text>;
  readonly version: Compact<u32>;
  readonly deprecationInfo: ItemDeprecationInfoV16;
}